requirements.txt
Django==1.11.5
django-ckeditor==5.3.0
django-js-asset==0.1.1
pep8==1.7.0
pytz==2017.2
